% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_MRC.R
\name{alpha_star_Intensity}
\alias{alpha_star_Intensity}
\title{alpha_star_Intensity}
\usage{
alpha_star_Intensity(Intensity, c0, c1, c2)
}
\arguments{
\item{Intensity}{a numeric vector or value, the precipitation intensity of interest}

\item{c0, c1, c2}{the parameters of the model}
}
\value{
a numeric value or a vector the same length as \code{length(Intensity)}
}
\description{
The quadratic model of the scaled \eqn{\alpha}. Given the parameters of the model \code{c0, c1, c2} and the 
precipitation intensity, it returns the value of scaled \code{alpha}.
}
\references{
Rupp, D. E., Keim, R. F., Ossiander, M., Brugnach, M., and Selker, J. S.:
Time scale and intensity dependency in multiplicative cascades for temporal rainfall disaggregation
Water Resources Research, 45,https://doi.org/10.1029/2008WR007321, 2009.
}
\author{
Kaltrina Maloku
}
