% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_data.R
\docType{data}
\name{data_test_MRC}
\alias{data_test_MRC}
\title{World Health Organization TB data}
\format{
## `data_test_MRC`
A data frame of two columns and 1577952 rows:
\describe{
  \item{obs}{Precipitation amounts}
  \item{date}{Dates}
}
}
\usage{
data_test_MRC
}
\description{
A fictive 10 minute resolution time series of precipitation
}
\keyword{datasets}
