% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_MRC.R
\name{estimate_params_MRC}
\alias{estimate_params_MRC}
\title{estimate_params_MRC}
\usage{
estimate_params_MRC(
  vecPrecip,
  vecDates,
  resVecPrecip = 10,
  by_season = T,
  threshold,
  threshold_alpha_int = 0.8,
  threshold_asymm = 0.8,
  aggLevels = c(40, 80, 160, 320, 640, 1280)
)
}
\arguments{
\item{vecPrecip}{a vector of observed precipitations}

\item{vecDates}{a vector of dates}

\item{resVecPrecip}{resolution of time series in minutes}

\item{by_season}{should the parameter estimation must be done on seasonal basis}

\item{threshold}{for weight discarding in all estimation procedure, in mm}

\item{threshold_alpha_int}{for discarding alpha estimated at small intensities, in mm}

\item{threshold_asymm}{for discarding weights of small intensities for estimation of asymmetry parameters, in mm}

\item{aggLevels}{aggregation levels in the cascade estimation procedure, in minutes}
}
\value{
to_do
}
\description{
Get parameters of MRC as estimated on observed weights
}
