% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_MRC.R
\name{fit_symm_beta_dist}
\alias{fit_symm_beta_dist}
\title{fit_symm_beta_dist}
\usage{
fit_symm_beta_dist(x)
}
\arguments{
\item{x}{a vector of weights}
}
\value{
estimated \eqn{\alpha} parameter
}
\description{
Fit symmetric beta distribution, with probability density function
\deqn{f(x;\alpha) = \frac{1}{B(\alpha, \alpha)} {x}^{(\alpha-1)}(1-x)^{(\alpha-1)}}
}
\details{
If the sample size is smaller than 10, return NA
}
\author{
Kaltrina Maloku
}
