% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_MRC.R
\name{get_alpha2}
\alias{get_alpha2}
\title{get_alpha2}
\usage{
get_alpha2(alpha, m)
}
\arguments{
\item{alpha}{a numeric or a  numeric vector of the parameter alpha of symmetric beta distribution}

\item{m}{a numeric or a  numeric vector of the mean of the distribution}
}
\value{
a numeric of the same length as \code{length(alpha)} with the parameter alpha2 of beta distribution
}
\description{
A function that estimates the parameter alpha2 of beta distribution 
given the parameter alpha of symmetric beta distribution and its mean
}
\author{
Kaltrina Maloku
}
