% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_MRC.R
\name{get_alpha_aggLevel}
\alias{get_alpha_aggLevel}
\title{get_alpha_aggLevel}
\usage{
get_alpha_aggLevel(alpha0, H, res_aggLevel)
}
\arguments{
\item{alpha0, H}{parameters of the model}

\item{res_aggLevel}{a numeric value, current aggregation level in minutes}
}
\value{
a numeric value, the value of the parameter \eqn{\alpha} given aggregation level
}
\description{
Model of alpha as function of temporal scale It returns the parameter \eqn{\alpha} of beta distribution 
given the parameters of the model \eqn{\alpha_{0}} and \eqn{H}
}
\references{
Rupp, D. E., Keim, R. F., Ossiander, M., Brugnach, M., and Selker, J. S.:
Time scale and intensity dependency in multiplicative cascades for temporal rainfall disaggregation
Water Resources Research, 45,https://doi.org/10.1029/2008WR007321, 2009.
}
\author{
Kaltrina Maloku
}
