% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_MRC.R
\name{get_alpha_intensity_1par}
\alias{get_alpha_intensity_1par}
\title{get_alpha_intensity_1par}
\usage{
get_alpha_intensity_1par(Intensity, I_min, I_max, K)
}
\arguments{
\item{Intensity}{a numeric vector or value, the precipitation intensity of interest}

\item{I_min}{for intensities smaller than this value the model is constant. Default value: 0.1 mm/h}

\item{I_max}{for intensities higher than this value the model is constant. Default value: 10 mm/h}

\item{K}{the parameter of the model}
}
\value{
a value or numeric vector of length \code{length(Intensity)} value
}
\description{
A model of the parameter \eqn{\alpha} as function of the precipitation intensity.
Its only parameter explains the degree of dependency to the intensity. 
The model is considered to be constant for intensities smaller than \code{I_min}, 
and higher than \code{I_max}.
}
\references{
my_paper
}
\author{
Kaltrina Maloku
}
