% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_MRC.R
\name{get_class_z_bin}
\alias{get_class_z_bin}
\title{get_class_z_bin}
\usage{
get_class_z_bin(z, bins_nb = 10)
}
\arguments{
\item{z}{a vector of z indices}

\item{bins_nb}{number of bins}
}
\value{
a numeric vector of the same length as x, indicating z index class
}
\description{
For a given z index, return the index of its bin class
}
\references{
My_paper
}
\author{
Kaltrina Maloku
}
