% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_MRC.R
\name{get_col_aggLevel}
\alias{get_col_aggLevel}
\title{get_col_aggLevel}
\usage{
get_col_aggLevel(aggLevel)
}
\arguments{
\item{aggLevel}{a vector providing the aggregation levels in minutes}
}
\value{
a vector of colors, same length as \code{length(aggLevel)}
}
\description{
Intermittency model, it returns the parameter \eqn{\alpha} of beta distribution 
given the parameters of the model \eqn{\alpha_{0}} and \eqn{H}
}
\author{
Kaltrina Maloku
}
