% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_MRC.R
\name{get_indices_center_1280}
\alias{get_indices_center_1280}
\title{get_indices_center_1280}
\usage{
get_indices_center_1280(nb_days)
}
\arguments{
\item{nb_days}{a numeric, number of days in the time series}
}
\value{
a numeric vector of indices
}
\description{
A function that helps extract 1280 minute "days" 
from a time series of 1440 minute days.  
Given the number of days, it returns the indices in 10 min data of a centered day at 12:00
under the assumption that the days are consecutive
}
\examples{
get_indices_center_1280(nb_days = 2)

}
\author{
Kaltrina Maloku
}
