% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_MRC.R
\name{get_intensity_class_id}
\alias{get_intensity_class_id}
\title{get_intensity_class_id}
\usage{
get_intensity_class_id(x, I_min = 0.1, res_aggLevel)
}
\arguments{
\item{x}{a vector of precipitation amounts}

\item{I_min}{Minimal observed precipitation in mm}

\item{res_aggLevel}{current aggregation level in minutes}
}
\value{
a vector of the same length as x indicating the class of each precip amount in mm/h
}
\description{
Get intensity id class given the precip amount
}
\references{
Rupp, D. E., Keim, R. F., Ossiander, M., Brugnach, M., and Selker, J. S.:
Time scale and intensity dependency in multiplicative cascades for temporal rainfall disaggregation
Water Resources Research, 45,https://doi.org/10.1029/2008WR007321, 2009.
}
\author{
Kaltrina Maloku
}
