% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_MRC.R
\name{get_mean_z}
\alias{get_mean_z}
\title{get_mean_z}
\usage{
get_mean_z(z, lambda)
}
\arguments{
\item{z}{a vector or a numeric value of z index}

\item{lambda}{parameter of the model}
}
\value{
a value or numeric vector of length \code{length(z)} value
}
\description{
Model of the mean of the cascade weights as function of asymmetry index z. 
Given the parameter of the model, \code{lambda}, and the asymmetry index z it returns 
modelled mean of the distribution.
}
\references{
my_paper
}
\author{
Kaltrina Maloku
}
