% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_MRC.R
\name{get_phi_z}
\alias{get_phi_z}
\title{get_phi_z}
\usage{
get_phi_z(z, nu)
}
\arguments{
\item{z}{a vector or a numeric value of z index}

\item{nu}{parameter of the model}
}
\value{
a value or numeric vector of length \code{length(z)}
}
\description{
Probability asymmetry ratio model, it returns the ratio
\eqn{\varphi = \frac{p_{01}{p_{01}+p{10}}}} given the precipitation asymmetry indez \code{z} 
and the parameter \code{nu}
}
\references{
my_paper
}
\author{
Kaltrina Maloku
}
