% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_MRC.R
\name{get_portion_of_0}
\alias{get_portion_of_0}
\title{get_portion_of_0}
\usage{
get_portion_of_0(x)
}
\arguments{
\item{x}{a vector of weights}
}
\value{
portion of weights equal to 0
}
\description{
Finds portion of cascade weights that are equal to 0
}
\details{
If the sample size is smaller than 10, return NA
}
\author{
Kaltrina Maloku
}
