% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_MRC.R
\name{get_portion_of_x}
\alias{get_portion_of_x}
\title{get_portion_of_1}
\usage{
get_portion_of_x(x)
}
\arguments{
\item{x}{a vector of weights}
}
\value{
portion of weights that are strictly greater than 0 and strictly smaller than 1
}
\description{
Finds portion of cascade weights that are strictly greater than 0 and strictly smaller than 1
}
\details{
If the sample size is smaller than 10, return NA
}
\author{
Kaltrina Maloku
}
