% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_MRC.R
\name{get_Px_Intensity}
\alias{get_Px_Intensity}
\title{get_Px_Intensity}
\usage{
get_Px_Intensity(Intensity, mu, Sigma)
}
\arguments{
\item{Intensity}{a numeric value, the precipitation intensity of interest}

\item{mu, Sigma}{parameters of the model}
}
\value{
a numeric value
}
\description{
Intermittency model, it returns the probability to divide precipitation intensity
given the parameters of the model \eqn{\mu} and \eqn{\sigma}
}
\references{
Rupp, D. E., Keim, R. F., Ossiander, M., Brugnach, M., and Selker, J. S.:
Time scale and intensity dependency in multiplicative cascades for temporal rainfall disaggregation
Water Resources Research, 45,https://doi.org/10.1029/2008WR007321, 2009.
}
\author{
Kaltrina Maloku
}
