% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_MRC.R
\name{get_Px_Intensity_aggLevel}
\alias{get_Px_Intensity_aggLevel}
\title{get_Px_Intensity_aggLevel}
\usage{
get_Px_Intensity_aggLevel(
  Intensity,
  res_aggLevel,
  a_mu,
  b_mu,
  a_sigma,
  b_sigma
)
}
\arguments{
\item{Intensity}{a numeric vector or value, the precipitation intensity of interest}

\item{res_aggLevel}{a numeric value, current aggregation level in minutes}

\item{a_mu, b_mu, a_sigma, b_sigma}{parameters of the model}
}
\value{
a numeric value or a vector the same length as \code{length(Intensity)}
}
\description{
Intermittency model, it returns the probability to divide precipitation intensity
given the parameters of the model \eqn{a_\mu,b_\mu,a_\sigma,b_\sigma} and 
current aggregation level
}
\references{
Rupp, D. E., Keim, R. F., Ossiander, M., Brugnach, M., and Selker, J. S.:
Time scale and intensity dependency in multiplicative cascades for temporal rainfall disaggregation
Water Resources Research, 45,https://doi.org/10.1029/2008WR007321, 2009.
}
\author{
Kaltrina Maloku
}
