% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_MRC.R
\name{get_z_index}
\alias{get_z_index}
\title{get_z_index}
\usage{
get_z_index(x)
}
\arguments{
\item{x}{a vector of precipitation amounts}
}
\value{
a vector of the same length as x, indicating asymmetry index z. 
It is calculated as \deqn{z_t =\frac{x_{t-1}+0.5 x_t}{x_{t-1}+x_t+x_{t+1}}}
}
\description{
Get class of a given precip amount
}
\references{
My_paper
}
\author{
Kaltrina Maloku
}
