% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_MRC.R
\name{sum_fixed_window}
\alias{sum_fixed_window}
\title{sum_fixed_window}
\usage{
sum_fixed_window(x, k)
}
\arguments{
\item{x}{a vector precipitations}

\item{k}{the length of fixed window `sum`}
}
\value{
vector of length \code{floor(length(x)/k)}, sum over the window of length \code{k}
}
\description{
Sum over a fixed window
}
\details{
Sums are assigned to the most left element of the window
}
\author{
Kaltrina Maloku
}
